fix: syscalls should return long number

+++ 1/initramfs/init.c
@@ -3,14 +3,14 @@
 
 #include "save_recover.h"
 int main(){
-	int random_n;
+	long random_n;
 	int pid = fork();
 	while(1) {
 		if (pid != 0) {
-			random_n = rand();
-			printf("Saving random number: %i\n", save(random_n));
+			random_n = (long) rand();
+			printf("Saving random number: %ld\n", save(random_n));
 		} else {
-			printf("Recovered random number: %i\n", recover());
+			printf("Recovered random number: %ld\n", recover());
 		}
 	}
 	return 0;
